requirements.txt
streamlit
pandas
xlsxwriter
Pillow
streamlit_calendar  
langcodes
python-i18n
